backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
jinja2
requests
pyyaml
pytest